--- client/components/login/LoginPage.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-import { observer } from 'mobx-react';
-import autoBindMethods from 'class-autobind-decorator';
-
-import LoginForm from './LoginForm';
-
-@autoBindMethods
-@observer
-class LoginPage extends Component {
-  render () {
-    return <LoginForm {...this.props} />;
-  }
-}
-
-export default LoginPage;
