--- client/consts.js
@@ -1,10 +1,0 @@
-const ACTIONS = {
-  'READ': 'READ',
-  'UNREAD': 'UNREAD',
-  'SKIP': 'SKIP',
-  'UNSKIP': 'UNSKIP',
-};
-
-export default {
-  ACTIONS,
-};
