+++ package.json
@@ -4,15 +4,21 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject",
     "heroku-postbuild": "yarn install && ./node_modules/webpack/bin/webpack.js -p",
-    "prod": " node server.js",
-    "start": "webpack-dev-server"
+    "prod": " node server.js"
   },
   "dependencies": {
-    "antd": "^3.1.0",
+    "@types/node": "^12.0.0",
+    "@types/react": "^16.8.16",
+    "@types/react-dom": "^16.8.4",
+    "@types/store": "^2.0.1",
+    "antd": "^3.16.6",
     "axios": "^0.17.1",
     "babel-core": "^6.21.0",
-    "babel-loader": "^7.1.2",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-polyfill": "^6.23.0",
     "babel-preset-latest": "^6.24.1",
@@ -22,7 +28,6 @@
     "classnames": "^2.2.5",
     "css-loader": "^0.28.8",
     "decorator-performance": "^0.1.0",
-    "eslint": "^4.14.0",
     "express": "^4.14.1",
     "html-webpack-plugin": "^2.26.0",
     "http-status-codes": "^1.3.0",
@@ -35,14 +40,25 @@
     "react": "16.2.0",
     "react-dom": "16.2.0",
     "react-router-dom": "^4.2.2",
+    "react-scripts": "^3.0.0",
     "store": "^2.0.12",
     "style-loader": "^0.19.1",
-    "webpack": "^3.10.0"
+    "typescript": "^3.4.5"
   },
   "devDependencies": {
-    "babel-eslint": "^8.1.2",
     "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-react": "^7.5.1",
-    "webpack-dev-server": "^2.10.0"
+    "eslint-plugin-react": "^7.5.1"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
