+++ src/components/resources/series/ComicsListColumns.tsx
