+++ src/consts.tsx
@@ -1,0 +1,13 @@
+export const ACTIONS = {
+  'READ': 'READ',
+  'UNREAD': 'UNREAD',
+  'SKIP': 'SKIP',
+  'UNSKIP': 'UNSKIP',
+};
+
+export const DATE_FORMAT = 'YYYY-MM-DD';
+
+export default {
+  ACTIONS,
+  DATE_FORMAT,
+};
